jekyll-filename-pretifier/requirements.txt
youtube-transcript-api
python-frontmatter
python-slugify
google-api-python-client
